REF: move to utils; AsyncioQueue inheritance -> composition
